Rename merlin.loader module to merlin.dataloader
Rename the main module to merlin.dataloader in order to remove confusion.
