+++ README.md
@@ -5,7 +5,7 @@
 ![GitHub License](https://img.shields.io/github/license/NVIDIA-Merlin/dataloader)
 [![Documentation](https://img.shields.io/badge/documentation-blue.svg)](https://nvidia-merlin.github.io/dataloader/main/README.html)
 
-The merlin-dataloader lets you quickly train recommender models for TensorFlow and PyTorch. It eliminates the biggest bottleneck in training recommender models, by providing GPU optimized dataloaders that read data directly into the GPU, and then do a 0-copy transfer to TensorFlow and PyTorch using [dlpack](https://github.com/dmlc/dlpack).
+The merlin-dataloader lets you quickly train recommender models for TensorFlow, PyTorch and JAX. It eliminates the biggest bottleneck in training recommender models, by providing GPU optimized dataloaders that read data directly into the GPU, and then do a 0-copy transfer to TensorFlow and PyTorch using [dlpack](https://github.com/dmlc/dlpack).
 
 
 The benefits of the Merlin Dataloader include:
@@ -20,7 +20,7 @@ Merlin-dataloader requires Python version 3.7+. Additionally, GPU support requir
 
 To install using Conda:
 ```
-conda install -c nvidia -c rapidsai -c numba -c conda-forge merlin-loader python=3.7 cudatoolkit=11.2
+conda install -c nvidia -c rapidsai -c numba -c conda-forge merlin-dataloader python=3.7 cudatoolkit=11.2
 ```
 
 To install from PyPi:
@@ -39,7 +39,7 @@ dataset = merlin.io.Dataset(PARQUET_FILE_PATHS, engine="parquet")
 
 # Create a Tensorflow dataloader from the dataset, loading 65K items
 # per batch
-from merlin.loader.tensorflow import Loader
+from merlin.dataloader.tensorflow import Loader
 loader = Loader(dataset, batch_size=65536)
 
 # Get a single batch of data. Inputs will be a dictionary of columnname
