+++ merlin/dataloader/ops/embeddings/torch_embedding_op.py
@@ -17,7 +17,7 @@
 import torch
 from torch.nn import Embedding
 
-from merlin.loader.ops.embeddings.embedding_op import (
+from merlin.dataloader.ops.embeddings.embedding_op import (
     EmbeddingOperator,
     MmapNumpyTorchEmbedding,
     NumpyEmbeddingOperator,
