+++ setup.cfg
@@ -35,10 +35,10 @@ ignore = E203,E402,W503
 [versioneer]
 VCS = git
 style = pep440
-versionfile_source = merlin/loader/_version.py
-versionfile_build = merlin/loader/_version.py
+versionfile_source = merlin/dataloader/_version.py
+versionfile_build = merlin/dataloader/_version.py
 tag_prefix = v
-parentdir_prefix = merlin-loader-
+parentdir_prefix = merlin-dataloader-
 
 [mypy]
 ignore_missing_imports = True
